pointerevents/pointerevent_touch-adjustment_click_target-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Touch-generated events should have the same target</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p>Touch letter 'O' below to run the test. If a "PASS" result appears the test passes, otherwise it fails</p>
<p><a href="#" id="link">Link</a> <span id="target">O</span></p>
<div id="log"></div>
<script>
async_test(t => {
    const link = document.getElementById('link');
    const target = document.getElementById('target');
    const expectedEventLog = ['pointerdown-link', 'touchstart-link', 'pointerup-link', 'touchend-link', 'click-link'];
    const eventLogRecorder = [];
    let pointerdownClientX, pointerdownClientY;

    const eventNames = ['touchstart', 'touchmove', 'touchend', 'pointerdown', 'pointermove', 'pointerup', 'click'];
    for (eventName of eventNames) {
        document.addEventListener(eventName, t.step_func(event => {
            // TouchEvent and PointerEvent should have the same un-adjusted coordinates.
            // click event should have coordinates adjusted to link element.
            const eventClientX = event.clientX || (event.touches.length > 0 ? event.touches[0].clientX : 0);
            const eventClientY = event.clientY || (event.touches.length > 0 ? event.touches[0].clientY : 0);
            if (event.type === 'pointerdown') {
                pointerdownClientX = eventClientX;
                pointerdownClientY = eventClientY;
                assert_equals(document.elementFromPoint(pointerdownClientX, pointerdownClientY), target,
                    'PointerEvent should have un-adjusted clientX/Y.');
            } else if (event.type === 'click') {
                assert_equals(document.elementFromPoint(eventClientX, eventClientY), link,
                    'click should have clientX/Y adjusted to link.');
            } else if (event.type !== 'touchend') {
                assert_equals(eventClientX, pointerdownClientX,
                    `${event.type} should have the same clientX as pointerdown.`);
                assert_equals(eventClientY, pointerdownClientY,
                    `${event.type} should have the same clientY as pointerdown.`);
            }

            // All events should have target adjusted to link.
            const targetName = event.target.id || event.target.nodeName || '[null]';
            eventLogRecorder.push(`${event.type}-${targetName}`);
            if (event.type === 'click') {
                assert_array_equals(eventLogRecorder, expectedEventLog,
                    `Expected: ${expectedEventLog}. Actual: ${eventLogRecorder}.`);
                t.done();
            }
        }));
    }
});
</script>
